+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased][unreleased]
 
+## [3.0.0][] - 2024-07-08
+
+### Major update
+
+- Realm caching support, now `CJS` realms depended on require are faster.
+- `CJS` type realms are receiving more native require construction now.
+- Options changes: `script` -> `realmOpts`, `run` -> `runOpts`.
+- New parameter `keepCache` for execute method.
+- Require function now part of Script.
+- Packages update.
+- Symbols rework.
+
 ## [2.2.0][] - 2024-05-07
 
 - Nix-Shell environment for minimal supported node version